create AuthenticationProvider for custom Authentication+update UserCo

+++ src/main/java/com/example/easybank/controller/AccountController.java
@@ -3,7 +3,6 @@ package com.example.easybank.controller;
 import com.example.easybank.model.Accounts;
 import com.example.easybank.repository.AccountsRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -16,7 +15,7 @@ public class AccountController {
 
 
     @GetMapping("/myAccount")
-    public Accounts getAccountDetails(@RequestParam int id) {
+    public Accounts getAccountDetails(@RequestParam long id) {
         Accounts accounts = accountRepository.findByCustomerId(id);
         if (accounts != null) {
             return accounts;
